refactor(agent-runtime): Plan节点拆分为4个单职责节点(W10/G-P3)
将plan.py(279行单体节点)拆为4个单一职责节点，对齐graph-engineering-product-spec
3.3.1矩阵定义，落实Hard Constraint「Plan节点必须拆分为更小单职责节点」。

节点流水线：decide_plan_mode → generate_plan → revise_manifest → compose_confirm
- decide_plan_mode: 纯判定，计算is_node_revise供下游使用
- generate_plan: 唯一LLM调用节点，生成/修改方案markdown
- revise_manifest: 条件LLM调用(node_revise时)，生成节点操作列表
- compose_confirm: 纯函数，组装确认消息+设置phase(SSOT for route_after_plan)

关键设计决策：
- 不使用LangGraph子图(子图丢失中间检查点)，保持扁平注册与W3一致
- revise_manifest在非node_revise时空穿透返回{}，无需条件路由
- route_after_plan简化为读phase字段，消除plan.py与builder.py的is_node_revise重复
- plan.py改为存根重新导出build_confirm_message，保持test_plan_summary.py兼容

state.py新增is_node_revise字段；builder.py路由目标名"plan"→"decide_plan_mode"/
"compose_confirm"；图编译+117测试通过(0新失败)。

spec文档graph-engineering-product-spec.md新增6A章节记录实施方案。
